requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
azure-storage-blob==12.23.1
python-multipart==0.0.20
numpy>=1.24.0
onnxruntime==1.19.0